Creating fngi programmer's manual
I've started working on the fngi programmer's manual with what little
bit I understand about fngi so far. So far I have documented how basic
values and functions work.

I Bob Forder assent to license this and all future contributions to
this project under the dual licenses of the UNLICENSE or MIT license
listed in the `UNLICENSE` and `README.md` files of this repository.